refactor: extracted Secure concern.
